requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
streamlit==1.39.0
pydantic==2.9.0
pydantic-settings==2.5.0

# LangGraph / LangChain
langgraph==0.2.35
langchain==0.3.0
langchain-openai==0.2.0

# OpenAI
openai==1.51.0
tiktoken==0.8.0
httpx==0.27.2

# Vector DB
chromadb==0.5.15

# BM25 fallback
rank-bm25==0.2.2
numpy==1.26.4

# PDF parsing (docling is an optional upgrade but pulls ~1GB of torch;
# the fallback to PyMuPDF + pdfplumber is used automatically when it is absent)
pymupdf==1.24.11
pdfplumber==0.11.4

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3

# Authentication
pyjwt==2.8.0
python-multipart==0.0.9

# Cost controls
pybreaker==1.0.1
redis==5.0.1

# Security
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
slowapi==0.1.9

# Observability
structlog==24.1.0
python-json-logger==2.0.7
prometheus-client==0.20.0

# Advanced features — Phase 9
faiss-cpu==1.7.4  # Vector index for semantic caching
# anthropic==0.20.0  # Uncomment for Claude model fallback (optional)

# HITL approval queue (celery is optional — only needed for distributed background approvals)
# celery==5.3.4  # Uncomment when adding background task workers

# RAG quality
flashrank==0.2.4
# sentence-transformers pulls ~2 GB of PyTorch. Only needed when
# FF_USE_CITATION_VALIDATION=true. Remove if you don't use that feature.
sentence-transformers==2.3.1

# OpenTelemetry tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Dev
python-dotenv==1.0.1
